pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pySimBlocks"
version = "0.1.0"
description = "A block-diagram simulation environment with Streamlit editor"
authors = [
    {name = "Antoine Alessandrini", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "streamlit",
    "pyyaml",
    "graphviz"
]

[project.scripts]
pysimblocks = "pySimBlocks.gui.launch:main"
pysimblocks-generate = "pySimBlocks.project.generate:main"
pysimblocks-update = "pySimBlocks.tools.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
examples = [
    "matplotlib",
    "control",
]

tests = [
    "control",
]
